requirements.txt
# ── Core Framework ────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# ── CLI ───────────────────────────────────────────────────────────────────────
click==8.1.7
rich==13.7.1                        # Beautiful terminal output

# ── Browser Automation ────────────────────────────────────────────────────────
playwright==1.44.0

# ── HTTP / API Testing ────────────────────────────────────────────────────────
httpx==0.27.0                       # Async HTTP client
aiohttp==3.9.5

# ── Config & Templating ───────────────────────────────────────────────────────
pyyaml==6.0.1
python-dotenv==1.0.1
jinja2==3.1.4                       # HTML report templating

# ── Database (async SQLAlchemy) ───────────────────────────────────────────────
sqlalchemy==2.0.30
aiosqlite==0.20.0                   # Async SQLite driver (MVP)
asyncpg==0.29.0                     # Async PostgreSQL driver (production)
alembic==1.13.1                     # DB migrations

# ── AI Providers ──────────────────────────────────────────────────────────────
openai==1.30.1
anthropic==0.26.1

# ── Utilities ─────────────────────────────────────────────────────────────────
python-multipart==0.0.9             # FastAPI file uploads (HAR files)
orjson==3.10.3                      # Fast JSON serialisation
tenacity==8.3.0                     # Retry logic with exponential backoff
structlog==24.2.0                   # Structured logging (JSON + console)
colorama==0.4.6                     # Windows-safe colour output

# ── Dev / Test ────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-playwright==0.4.4
httpx==0.27.0                       # Also used in tests as async test client
